Remove modules
Modules previously served as a basic level of abstraction which is now properly covered by the commands package. The two independent modules will simply be brought into the base package for the time being
